# Shorter blocks and consistent file