Lepton format allows to write blocks of standard length (65536 bytes) shorter, with 1-byte header instead of 3-bytes. Additionally, current implementation writes down encoded blocks into the file in the order of their receiving from different threads - that leads to different output files across runs.

This PR makes the output file shorter and the order of blocks in it predefined by carouseling. Performance is increased by ~3 % due to omission of buffer size check at each stream bit pushing.

Current best of 3 runs:
```
2024-11-14T17:00:03.681Z INFO  [lepton_jpeg::structs::lepton_file_writer] compressing to Lepton format
2024-11-14T17:00:04.152Z INFO  [lepton_jpeg::structs::lepton_file_writer] Number of threads: 8
2024-11-14T17:00:05.668Z INFO  [lepton_jpeg::structs::lepton_file_writer] worker threads 10212ms of CPU time in 1514ms of wall time
2024-11-14T17:00:05.668Z INFO  [lepton_jpeg::structs::lepton_file_writer] decompressing to verify contents
2024-11-14T17:00:07.276Z INFO  [lepton_jpeg_util] compressed input 22171278, output 17324596 bytes (compression = 28.0%)
2024-11-14T17:00:07.276Z INFO  [lepton_jpeg_util] Main thread CPU: 3595ms, Worker thread CPU: 21306 ms, walltime: 3595 ms

 Performance counter stats for 'taskset -c 10 nice -n -20 target/release/lepton_jpeg_util images/img_52MP_7k.jpg images/img_52MP_7k2.lep':

       855 564 085      cache-references                                                        (42,07%)
        86 316 838      cache-misses                     #   10,09% of all cache refs           (42,21%)
    16 099 617 774      cycles                                                                  (42,31%)
       806 699 300      ic_fetch_stall.ic_stall_back_pressure                                        (42,10%)
     1 061 388 071      stalled-cycles-frontend          #    6,59% frontend cycles idle        (41,84%)
    40 125 811 710      instructions                     #    2,49  insn per cycle            
                                                  #    0,03  stalled cycles per insn     (41,72%)
     4 707 746 288      branch-instructions                                                     (41,69%)
       165 544 725      branch-misses                    #    3,52% of all branches             (41,64%)
     5 309 260 356      ic_fetch_stall.ic_stall_any                                             (41,56%)
        43 294 231      ic_fetch_stall.ic_stall_dq_empty                                        (41,80%)
        73 167 746      l2_cache_misses_from_ic_miss                                            (42,11%)
     2 188 259 483      l2_latency.l2_cycles_waiting_on_fills                                        (42,11%)
           183 928      faults                                                                
                 1      migrations                                                            

       3,628985546 seconds time elapsed

       3,289304000 seconds user
       0,336724000 seconds sys
```
This PR  https://github.com/microsoft/lepton_jpeg_rust/pull/112/commits/c7e0f7cea74909140464b08c80b1004f1f0580b6 best of 3 runs:
```
2024-11-14T21:13:33.559Z INFO  [lepton_jpeg::structs::lepton_file_writer] compressing to Lepton format
2024-11-14T21:13:34.039Z INFO  [lepton_jpeg::structs::lepton_file_writer] Number of threads: 8
2024-11-14T21:13:35.420Z INFO  [lepton_jpeg::structs::lepton_file_writer] worker threads 9299ms of CPU time in 1379ms of wall time
2024-11-14T21:13:35.420Z INFO  [lepton_jpeg::structs::lepton_file_writer] decompressing to verify contents
2024-11-14T21:13:37.018Z INFO  [lepton_jpeg_util] compressed input 22171278, output 17324076 bytes (compression = 28.0%)
2024-11-14T21:13:37.018Z INFO  [lepton_jpeg_util] Main thread CPU: 3458ms, Worker thread CPU: 20273 ms, walltime: 3458 ms

 Performance counter stats for 'taskset -c 10 nice -n -20 target/release/lepton_jpeg_util images/img_52MP_7k.jpg images/img_52MP_7k2.lep':

       834 601 112      cache-references                                                        (42,12%)
        64 932 273      cache-misses                     #    7,78% of all cache refs           (42,04%)
    15 586 303 408      cycles                                                                  (42,04%)
       838 269 598      ic_fetch_stall.ic_stall_back_pressure                                        (41,98%)
     1 061 874 437      stalled-cycles-frontend          #    6,81% frontend cycles idle        (41,79%)
    38 330 557 068      instructions                     #    2,46  insn per cycle            
                                                  #    0,03  stalled cycles per insn     (41,68%)
     4 353 138 372      branch-instructions                                                     (41,68%)
       163 333 881      branch-misses                    #    3,75% of all branches             (41,72%)
     5 226 409 949      ic_fetch_stall.ic_stall_any                                             (41,86%)
        39 792 649      ic_fetch_stall.ic_stall_dq_empty                                        (41,92%)
        53 084 247      l2_cache_misses_from_ic_miss                                            (41,96%)
     2 066 007 058      l2_latency.l2_cycles_waiting_on_fills                                        (42,16%)
           183 967      faults                                                                
                 1      migrations                                                            

       3,492682793 seconds time elapsed

       3,175743000 seconds user
       0,313777000 seconds sys
```