@Melirius 2024-11-14
> Better to separate out the multiplexer changes and work on the bitwriter changes separately.

Separate into 2 PRs? Or leave like this?