@mcroomp 2024-11-15
> > Better to separate out the multiplexer changes and work on the bitwriter changes separately.
> 
> Separate into 2 PRs? Or leave like this?

I think in 2 PRs would be better since they aren't really related. Trying to be extra careful with the write codepath since corruptions on the write side are much more dangerous. Thanks!