@Melirius 2024-11-15
This version https://github.com/microsoft/lepton_jpeg_rust/pull/112/commits/f655f27586741c380a6d488643e7dec791882afa is slower but still faster than main branch:
```
2024-11-15T11:35:16.838Z INFO  [lepton_jpeg::structs::lepton_file_writer] compressing to Lepton format
2024-11-15T11:35:17.308Z INFO  [lepton_jpeg::structs::lepton_file_writer] Number of threads: 8
2024-11-15T11:35:18.756Z INFO  [lepton_jpeg::structs::lepton_file_writer] worker threads 9592ms of CPU time in 1447ms of wall time
2024-11-15T11:35:18.756Z INFO  [lepton_jpeg::structs::lepton_file_writer] decompressing to verify contents
2024-11-15T11:35:20.364Z INFO  [lepton_jpeg_util] compressed input 22171278, output 17324076 bytes (compression = 28.0%)
2024-11-15T11:35:20.364Z INFO  [lepton_jpeg_util] Main thread CPU: 3526ms, Worker thread CPU: 20654 ms, walltime: 3526 ms

 Performance counter stats for 'taskset -c 10 nice -n -20 target/release/lepton_jpeg_util images/img_52MP_7k.jpg images/img_52MP_7k2.lep':

       867 203 526      cache-references                                                        (42,48%)
        84 794 880      cache-misses                     #    9,78% of all cache refs           (42,32%)
    15 852 334 145      cycles                                                                  (42,23%)
     1 013 465 354      ic_fetch_stall.ic_stall_back_pressure                                        (42,01%)
     1 069 301 834      stalled-cycles-frontend          #    6,75% frontend cycles idle        (41,84%)
    39 172 164 846      instructions                     #    2,47  insn per cycle            
                                                  #    0,03  stalled cycles per insn     (41,67%)
     4 396 742 683      branch-instructions                                                     (41,58%)
       165 106 663      branch-misses                    #    3,76% of all branches             (41,46%)
     5 327 122 036      ic_fetch_stall.ic_stall_any                                             (41,54%)
        38 628 152      ic_fetch_stall.ic_stall_dq_empty                                        (41,71%)
        71 604 981      l2_cache_misses_from_ic_miss                                            (41,92%)
     2 071 017 680      l2_latency.l2_cycles_waiting_on_fills                                        (42,27%)
           183 950      faults                                                                
                 1      migrations                                                            

       3,561160101 seconds time elapsed

       3,216183000 seconds user
       0,342806000 seconds sys
```